+++ src/modules/order/order.service.ts
@@ -637,199 +637,6 @@ export class OrderService extends BaseUserService<Order> {
     return /[\u0600-\u06FF]/.test(text);
   }
 
-  async returnOrder(order_id: string, req: ReturnOrderRequest) {
-    const { returned_shipment_products } = req;
-    const returned_shipment_products_id = returned_shipment_products.map(
-      (p) => p.shipment_product_id,
-    );
-
-    // check if the order exists
-    const order = await this.orderRepository.findOne({
-      where: { id: order_id },
-    });
-    if (!order) throw new NotFoundException('order not found');
-
-    // check if the order belongs to the current user
-    if (order.user_id !== this.currentUser.id) {
-      throw new BadRequestException('you are not allowed to return this order');
-    }
-
-    // check if the order is delivered
-    const shipment = await this.shipmentRepository.findOne({
-      where: { order_id },
-      relations: ['warehouse'],
-    });
-    if (!shipment) throw new NotFoundException('shipment not found');
-
-    if (
-      shipment.status !== ShipmentStatusEnum.DELIVERED &&
-      shipment.status !== ShipmentStatusEnum.COMPLETED &&
-      shipment.status !== ShipmentStatusEnum.CANCELED
-    ) {
-      throw new BadRequestException(
-        'you can not return this order that is not delivered yet',
-      );
-    }
-
-    // check if the return products IDs are valid and belongs to the order
-    const shipmentProducts = await this.shipmentProductRepository.find({
-      where: {
-        shipment_id: shipment.id,
-        id: In(returned_shipment_products_id),
-      },
-    });
-
-    if (shipmentProducts.length !== returned_shipment_products.length) {
-      throw new BadRequestException('invalid products IDs');
-    }
-
-    const canNotReturnProducts = shipmentProducts.find(
-      (p) => p.can_return === false,
-    );
-    if (canNotReturnProducts) {
-      throw new BadRequestException(
-        `you can't return this product alread returned ${JSON.stringify(
-          canNotReturnProducts,
-        )}`,
-      );
-    }
-
-    // check if the return products quantities are valid based on the shipment products
-    const invalidQuantityForProducts = returned_shipment_products.filter(
-      (p) => {
-        const product = shipmentProducts.find(
-          (sp) => sp.id === p.shipment_product_id,
-        );
-        return product.quantity < p.quantity;
-      },
-    );
-
-    if (invalidQuantityForProducts.length > 0) {
-      throw new BadRequestException(
-        `invalid quantity for products ${JSON.stringify(
-          invalidQuantityForProducts,
-        )}`,
-      );
-    }
-
-    // check if the return products reasons IDs are valid
-    const returnedProductsReasons =
-      await this.returnProductReasonRepository.find({
-        where: {
-          id: In(returned_shipment_products.map((p) => p.reason_id)),
-        },
-      });
-
-    if (
-      returnedProductsReasons.length !==
-      new Set(returned_shipment_products.map((p) => p.reason_id)).size
-    ) {
-      throw new BadRequestException(`invalid return reasons IDs`);
-    }
-
-    const returnOrder = await this.ReturnOrderRepository.create({
-      status: ReturnOrderStatus.PENDING,
-      order,
-      returnOrderProducts: req.returned_shipment_products,
-      customer_note: req.customer_note,
-    });
-
-    await this.shipmentProductRepository.update(
-      {
-        id: In(returned_shipment_products_id),
-      },
-      {
-        can_return: false,
-      },
-    );
-
-    const savedReturnOrder = await this.ReturnOrderRepository.save(returnOrder);
-    await this.orderGateway.notifyReturnOrder({
-      to_rooms: ['admin'],
-      body: {
-        client: this.currentUser,
-        driver: null,
-        order,
-        return_order: savedReturnOrder,
-        warehouse: shipment.warehouse,
-      },
-    });
-
-    return savedReturnOrder;
-  }
-
-  // this method is used by the admin to update the return order status
-  async updateReturnOrderStatus(
-    return_order_id: string,
-    req: UpdateReturnOrderStatusRequest,
-  ) {
-    const returnOrder = await this.ReturnOrderRepository.findOne({
-      where: { id: return_order_id },
-      relations: ['order', 'order.user'],
-    });
-
-    if (!returnOrder) throw new NotFoundException('return order not found');
-
-    const { return_order_products, admin_note, driver_id, status } = req;
-    const returned_products_id = return_order_products.map(
-      (p) => p.return_order_product_id,
-    );
-
-    // check if the return order products IDs are valid
-    const returnOrderProducts = await this.returnOrderProductRepository.find({
-      where: {
-        id: In(returned_products_id),
-        return_order_id,
-      },
-    });
-
-    if (returnOrderProducts.length !== return_order_products.length) {
-      throw new BadRequestException('invalid return order products IDs');
-    }
-
-    let driver: Driver = null;
-    if (driver_id) {
-      driver = await this.driverRepository.findOne({
-        where: { id: driver_id },
-      });
-      if (!driver) throw new BadRequestException('driver not found');
-    }
-
-    // mapped the return order products with the updated status
-    const mappedUpdatedReturnOrderProducts = returnOrderProducts.map((p) => {
-      const product = return_order_products.find(
-        (rp) => rp.return_order_product_id === p.id,
-      );
-      return {
-        ...p,
-        status: product.status,
-      };
-    });
-
-    await this.returnOrderProductRepository.save(
-      mappedUpdatedReturnOrderProducts,
-    );
-    const savedReturnOrder = await this.ReturnOrderRepository.save(returnOrder);
-
-    const shipment = await this.shipmentRepository.findOne({
-      where: { order_id: returnOrder.order_id },
-      relations: ['warehouse'],
-    });
-
-    await this.orderGateway.notifyReturnOrder({
-      to_rooms: [returnOrder.order.user.id, driver?.id],
-      body: {
-        client: returnOrder.order.user,
-        driver: driver,
-        order: returnOrder.order,
-        return_order: savedReturnOrder,
-        warehouse: shipment.warehouse,
-      },
-    });
-
-    return savedReturnOrder;
-  }
-
   async broadcastOrderDrivers(order_id: string) {
     const order = await this.orderRepository.findOne({
       where: { id: order_id },
